% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{omics.network}
\alias{omics.network}
\title{Omics Network Object}
\description{
Omics Network Object

Omics Network Object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{omics.network$new()}}
\item \href{#method-peek}{\code{omics.network$peek()}}
\item \href{#method-nodes}{\code{omics.network$nodes()}}
\item \href{#method-nodes.map}{\code{omics.network$nodes.map()}}
\item \href{#method-nodes.annotate}{\code{omics.network$nodes.annotate()}}
\item \href{#method-nodes.attributes}{\code{omics.network$nodes.attributes()}}
\item \href{#method-nodes.filter}{\code{omics.network$nodes.filter()}}
\item \href{#method-nodes.update}{\code{omics.network$nodes.update()}}
\item \href{#method-nodes.neighbors}{\code{omics.network$nodes.neighbors()}}
\item \href{#method-edges}{\code{omics.network$edges()}}
\item \href{#method-edges.map}{\code{omics.network$edges.map()}}
\item \href{#method-edges.annotate}{\code{omics.network$edges.annotate()}}
\item \href{#method-edges.attributes}{\code{omics.network$edges.attributes()}}
\item \href{#method-edges.filter}{\code{omics.network$edges.filter()}}
\item \href{#method-edges.update}{\code{omics.network$edges.update()}}
\item \href{#method-graph.update}{\code{omics.network$graph.update()}}
\item \href{#method-graph.on.change}{\code{omics.network$graph.on.change()}}
\item \href{#method-graph.delete.nodes}{\code{omics.network$graph.delete.nodes()}}
\item \href{#method-graph.delete.edges}{\code{omics.network$graph.delete.edges()}}
\item \href{#method-graph.simplify}{\code{omics.network$graph.simplify()}}
\item \href{#method-graph.delete.isolates}{\code{omics.network$graph.delete.isolates()}}
\item \href{#method-graph.subset.nodes}{\code{omics.network$graph.subset.nodes()}}
\item \href{#method-graph.subset.edges}{\code{omics.network$graph.subset.edges()}}
\item \href{#method-plot}{\code{omics.network$plot()}}
\item \href{#method-clone}{\code{omics.network$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$new(ig, node.fn = list(), edge.fn = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-peek"></a>}}
\if{latex}{\out{\hypertarget{method-peek}{}}}
\subsection{Method \code{peek()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$peek(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodes"></a>}}
\if{latex}{\out{\hypertarget{method-nodes}{}}}
\subsection{Method \code{nodes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$nodes(attr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodes.map"></a>}}
\if{latex}{\out{\hypertarget{method-nodes.map}{}}}
\subsection{Method \code{nodes.map()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$nodes.map(ids, attr_1, attr_2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodes.annotate"></a>}}
\if{latex}{\out{\hypertarget{method-nodes.annotate}{}}}
\subsection{Method \code{nodes.annotate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$nodes.annotate(values, attr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodes.attributes"></a>}}
\if{latex}{\out{\hypertarget{method-nodes.attributes}{}}}
\subsection{Method \code{nodes.attributes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$nodes.attributes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodes.filter"></a>}}
\if{latex}{\out{\hypertarget{method-nodes.filter}{}}}
\subsection{Method \code{nodes.filter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$nodes.filter(conditions, attr = "name")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodes.update"></a>}}
\if{latex}{\out{\hypertarget{method-nodes.update}{}}}
\subsection{Method \code{nodes.update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$nodes.update()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nodes.neighbors"></a>}}
\if{latex}{\out{\hypertarget{method-nodes.neighbors}{}}}
\subsection{Method \code{nodes.neighbors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$nodes.neighbors(
  ids,
  attr = "name",
  degree = 1,
  neighbors.only = TRUE,
  return.names = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-edges"></a>}}
\if{latex}{\out{\hypertarget{method-edges}{}}}
\subsection{Method \code{edges()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$edges(attr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-edges.map"></a>}}
\if{latex}{\out{\hypertarget{method-edges.map}{}}}
\subsection{Method \code{edges.map()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$edges.map(ids, attr_1, attr_2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-edges.annotate"></a>}}
\if{latex}{\out{\hypertarget{method-edges.annotate}{}}}
\subsection{Method \code{edges.annotate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$edges.annotate(values, attr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-edges.attributes"></a>}}
\if{latex}{\out{\hypertarget{method-edges.attributes}{}}}
\subsection{Method \code{edges.attributes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$edges.attributes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-edges.filter"></a>}}
\if{latex}{\out{\hypertarget{method-edges.filter}{}}}
\subsection{Method \code{edges.filter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$edges.filter(conditions, attr = "name")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-edges.update"></a>}}
\if{latex}{\out{\hypertarget{method-edges.update}{}}}
\subsection{Method \code{edges.update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$edges.update()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-graph.update"></a>}}
\if{latex}{\out{\hypertarget{method-graph.update}{}}}
\subsection{Method \code{graph.update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$graph.update()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-graph.on.change"></a>}}
\if{latex}{\out{\hypertarget{method-graph.on.change}{}}}
\subsection{Method \code{graph.on.change()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$graph.on.change(update)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-graph.delete.nodes"></a>}}
\if{latex}{\out{\hypertarget{method-graph.delete.nodes}{}}}
\subsection{Method \code{graph.delete.nodes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$graph.delete.nodes(ids, attr = "name", update = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-graph.delete.edges"></a>}}
\if{latex}{\out{\hypertarget{method-graph.delete.edges}{}}}
\subsection{Method \code{graph.delete.edges()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$graph.delete.edges(ids, attr = "name", update = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-graph.simplify"></a>}}
\if{latex}{\out{\hypertarget{method-graph.simplify}{}}}
\subsection{Method \code{graph.simplify()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$graph.simplify(
  remove.multiple = TRUE,
  remove.loops = TRUE,
  update = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-graph.delete.isolates"></a>}}
\if{latex}{\out{\hypertarget{method-graph.delete.isolates}{}}}
\subsection{Method \code{graph.delete.isolates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$graph.delete.isolates(update = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-graph.subset.nodes"></a>}}
\if{latex}{\out{\hypertarget{method-graph.subset.nodes}{}}}
\subsection{Method \code{graph.subset.nodes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$graph.subset.nodes(ids, attr = "name", degree = 0, update = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-graph.subset.edges"></a>}}
\if{latex}{\out{\hypertarget{method-graph.subset.edges}{}}}
\subsection{Method \code{graph.subset.edges()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$graph.subset.edges(ids, update = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$plot(
  vertex.size = 4,
  vertex.color = "#EEEEEE",
  vertex.frame.color = "#000040",
  vertex.label = NA,
  vertex.label.family = "Helvetica",
  vertex.label.color = "#000000",
  vertex.label.font = 2,
  vertex.label.cex = 0.75,
  vertex.label.dist = 0,
  edge.width = 1,
  edge.color = "#3D3D3D",
  edge.arrow.size = 0.2,
  no.margin = TRUE,
  layout = NULL,
  seed = 1,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{omics.network$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
